Move LexActiveContext to its own file
