+++ Libraries/src/Amazon.Lambda.LexEvents/LexEvent.cs
@@ -92,43 +92,6 @@ namespace Amazon.Lambda.LexEvents
         }
 
         /// <summary>
-        /// One or more contexts that are active during this turn of a conversation with the user.
-        /// </summary>
-        public class LexActiveContext
-        {
-            /// <summary>
-            /// The length of time or number of turns in the conversation with the user that the context remains active.
-            /// </summary>
-            public TimeToLive TimeToLive { get; set; }
-
-            /// <summary>
-            /// The name of the context.
-            /// </summary>
-            public string Name { get; set; }
-
-            /// <summary>
-            /// A list of key/value pairs the contains the name and value of the slots from the intent that activated the context.
-            /// </summary>
-            public IDictionary<string, string> Parameters { get; set; }
-        }
-
-        /// <summary>
-        /// The length of time or number of turns in the conversation with the user that the context remains active.
-        /// </summary>
-        public class TimeToLive
-        {
-            /// <summary>
-            /// The length of time that the context remains active.
-            /// </summary>
-            public int TimeToLiveInSeconds { get; set; }
-
-            /// <summary>
-            /// The number of turns in the conversation with the user that the context remains active.
-            /// </summary>
-            public int TurnsToLive { get; set; }
-        }
-
-        /// <summary>
         /// The class representing the current intent for the Lambda function to process.
         /// </summary>
         public class LexCurrentIntent
